+++ backend/package.json
@@ -29,13 +29,14 @@
   },
   "dependencies": {
     "@vscode-logging/logger": "^2.0.0",
+    "axios": "^1.6.8",
     "bcrypt": "^5.1.1",
-    "axios": "^1.6.7",
     "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
     "express": "^4.18.3",
-    "mongodb": "^6.5.0",
     "express-session": "^1.18.0",
     "jsonwebtoken": "^9.0.2",
+    "mongodb": "^6.5.0",
     "mongoose": "^8.2.1",
     "passport": "^0.7.0",
     "passport-local": "^1.0.0"
