+++ backend/src/controllers/user.ts
@@ -1,9 +1,7 @@
 import { Request, Response, response } from "express";
 import bcrypt from "bcrypt";
-const loginRouter = require("express").Router();
 const usersRouter = require("express").Router();
 const User = require("../models/user");
-const jwt = require("jsonwebtoken");
 
 usersRouter.post("/", async (request: Request, response: Response) => {
   const { username, name, password } = request.body;
@@ -20,30 +18,9 @@ usersRouter.post("/", async (request: Request, response: Response) => {
   response.json(savedUser);
 });
 
-loginRouter.post('/', async (request: Request, response: Response) => {
-  const { username, password } = request.body;
-  
-  const user = await User.findOne({ username });
-  const passwordCorrect = user === null
-    ? false
-    : await bcrypt.compare(password, user.passwordHash);
+/*usersRouter.get("/", async (_request: Request, response: Response) => {
+  const users = await User.find({});
+  response.json(users);
+});*/
 
-  if (!(user && passwordCorrect)) {
-    return response.status(401).json({
-      error: "invalid username or password"
-    })
-  }
-
-  const userForToken = {
-    username: user.username,
-    id: user._id,
-  }
-
-  const token = jwt.sign(userForToken, process.env.SECRET);
-
-  response
-    .status(200)
-    .send({ token, username: user.username });
-});
-
-module.exports = { loginRouter, usersRouter };
+export default usersRouter;
